pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Anaconda Training Project"
authors = [
    {name = "Ganesh",email = "kganeshe2@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "polars>=0.20.0",
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "python-dateutil>=2.8.0",
    "pandas>=2.2.0,<3.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
